exercie 2 finished, interesting use of BigDecimal and its integrated functions to make things works as expected

+++ src/serie5/PaymentMethod.java
@@ -1,6 +1,8 @@
 package serie5;
 
 
+import java.math.BigDecimal;
+
 interface PaymentMethod {
-    void pay(double amount, BankAccount destinationAccount);
+    void pay(BigDecimal amount, BankAccount destinationAccount);
 }
